fix contactsMessengerPicker and test
