+++ src/components/common/contactsMessengerPicker/ContactsMessengerPicker.module.css
@@ -13,7 +13,6 @@
 .statusWrapper {
   position: absolute;
   width: 100%;
-  height: 50px;
   border-radius: 12px;
   border: solid 1px var(--border-color);
   z-index: 1;
@@ -22,10 +21,6 @@
 
 .statusWrapper:focus-within {
   border-color: var(--primary-color);
-}
-
-.openedStatusWrapper {
-  height: 210px;
 }
 
 .errorStatusWrapper {
@@ -102,11 +97,6 @@
   padding-inline-start: 0px;
   padding: 13px 15px;
 }
-
-.hiddenList {
-  visibility: hidden;
-}
-
 .item {
   display: flex;
   align-items: center;
